requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
bcrypt==4.2.1
python-dotenv==1.0.1
WTForms==3.2.1
email-validator==2.2.0
